requirements.txt
aiohttp>=3.9,<4.0
python-dotenv>=1.0
SQLAlchemy>=2.0,<3.0
mysql-connector-python>=8.0
